+++ src/command.games.user.js
@@ -1,20 +1,20 @@
 // ==UserScript==
 // @name         command.games
 // @namespace    https://github.com/davidtorosyan/command.games
-// @version      1.7.0
+// @version      1.8.0
 // @description  improve dominion.games
 // @author       David Torosyan
 // @match        https://dominion.games/*
 // @match        https://dominionrandomizer.com/*
 // @require      https://code.jquery.com/jquery-3.4.1.min.js
-// @require      https://github.com/davidtorosyan/command.games/raw/monkeymaster-v1.3.1/src/monkeymaster/monkeymaster.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/common.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/style.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/background.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/random.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/export.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/kingdom.js
-// @require      https://github.com/davidtorosyan/command.games/raw/v1.7.0/src/modules/main.js
+// @require      https://github.com/davidtorosyan/command.games/raw/monkeymaster-v2.0.0/src/monkeymaster/monkeymaster.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/common.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/style.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/background.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/random.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/export.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/kingdom.js
+// @require      https://github.com/davidtorosyan/command.games/raw/v1.8.0/src/modules/main.js
 // @grant        GM_setValue
 // @grant        GM_getValue
 // @grant        GM_addValueChangeListener
@@ -31,6 +31,4 @@
     1. All files under the modules directory must be included.
     2. The first three files must be jQuery, monkeymaster, and common (in that order).
     3. The last file must be main.
-
-    Other than those rules, there should be no dependencies between modules.
 */