+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2020-09-29
+
+### Added
+
+- [FEATURE #13](https://github.com/davidtorosyan/command.games/issues/13): partial kingdom randomization!
+
+### Fixed
+
+- Better error handling, so the Random button is now reset when there's an issue.
+
 ## [1.7.0] - 2020-09-28
 
 ### Added
@@ -102,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - command.games userscript and monkeymaster library
 - extends https://dominion.games/ to have a "Random!" button during card selection
 
-[unreleased]: https://github.com/davidtorosyan/command.games/compare/v1.7.0...HEAD
+[unreleased]: https://github.com/davidtorosyan/command.games/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/davidtorosyan/command.games/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/davidtorosyan/command.games/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/davidtorosyan/command.games/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/davidtorosyan/command.games/compare/v1.4.1...v1.5.0